+++ src/config/constant/constant.go
@@ -1,3 +1,0 @@
-package constant
-
-const EXAMPLE = "example"