+++ src/database/mysql/mysql_test.go
@@ -1,7 +1,0 @@
-package mysql
-
-import "testing"
-
-func TestConnect(t *testing.T) {
-	panic("implement me")
-}