+++ src/delivery/rest/gingonic/example_v1_test.go
@@ -1,7 +1,0 @@
-package gingonic_test
-
-import "testing"
-
-func TestHandler_ExampleGet(t *testing.T) {
-	panic("implement me")
-}