+++ src/delivery/rest/gingonic/routes.go
@@ -1,6 +1,0 @@
-package gingonic
-
-// Routes registers gin-gonic routes.
-func (h *Handler) Routes() {
-	panic("implement me")
-}