+++ src/delivery/rest/gingonic/routes_test.go
@@ -1,7 +1,0 @@
-package gingonic_test
-
-import "testing"
-
-func TestHandler_Routes(t *testing.T) {
-	panic("implement me")
-}