+++ src/entity/example.go
@@ -1,9 +1,0 @@
-package entity
-
-import "time"
-
-type Example struct {
-	Id        int       `json:"id"`
-	Name      string    `json:"name"`
-	CreatedAt time.Time `json:"created_at"`
-}