+++ src/library/printme/printme.go
@@ -1,6 +1,0 @@
-package printme
-
-// PrintMe prints the given text.
-func PrintMe(text string) {
-	panic("implement me")
-}