+++ src/library/printme/printme_test.go
@@ -1,7 +1,0 @@
-package printme_test
-
-import "testing"
-
-func TestPrintMe(t *testing.T) {
-	panic("implement me")
-}