+++ src/service/service_test.go
@@ -1,9 +1,0 @@
-package service_test
-
-import (
-	"testing"
-)
-
-func TestNew(t *testing.T) {
-	panic("implement me")
-}