+++ structure/src/adapter/sqlite/sqlite.go
@@ -1,0 +1,11 @@
+package sqlite
+
+import (
+	"database/sql"
+)
+
+// New
+
+func NewSQLite(connString string) (*sql.DB, error) {
+	panic("implement me")
+}