+++ structure/src/config/env/env.go
@@ -1,0 +1,7 @@
+package env
+
+// Parse
+
+func Parse(scan interface{}) error {
+	panic("implement me")
+}