+++ structure/src/entity/example/base_model.go
@@ -1,0 +1,5 @@
+package example
+
+type BaseModel struct {
+	ID uint64 `json:"id"`
+}