+++ structure/src/library/example/printme.go
@@ -1,0 +1,5 @@
+package example
+
+func Example() {
+	panic("implement me")
+}