+++ structure/src/repository/example/contract.go
@@ -1,0 +1,10 @@
+package example
+
+import (
+	"context"
+	"github.com/dalikewara/ayapingping-go/v3/structure/src/entity/example"
+)
+
+type Contract interface {
+	ExecuteCtx(ctx context.Context, id string) (*example.BaseModel, error)
+}