+++ structure/src/service/rest/contract.go
@@ -1,0 +1,5 @@
+package rest
+
+type Contract interface {
+	Serve()
+}