+++ structure/src/service/rest/route.go
@@ -1,0 +1,3 @@
+package rest
+
+const RoutePing = "/ping"